Authentication

+++ createAdventure/accommodation/api/views.py
@@ -2,10 +2,13 @@ from django.core.checks import messages
 from django.shortcuts import render
 from django.http import HttpResponse, JsonResponse, Http404
 from requests import Response
+from rest_framework.authentication import TokenAuthentication
 from rest_framework.decorators import api_view
 from rest_framework.parsers import JSONParser
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
 
+from account.models import update_auth_token
 from .models import Accommodation as AccommodationModel
 from .serializer import AccommodationSerializer
 from rest_framework import status, generics
@@ -61,28 +64,45 @@ class listOfAccommodations(APIView):
     List all accommodations, or create a new accommodation.
     """
 
+    # authentication_classes = [TokenAuthentication]
+    permission_classes = [IsAuthenticated]
+
     schema = AccommodationSchema()
 
     def get(self, request):
-        accommodation = AccommodationModel.objects.all()
-        if accommodation.count() > 0:
-            serializer = AccommodationSerializer(accommodation, many=True)
-            return JsonResponse(serializer.data, status=status.HTTP_200_OK, safe=False)
+        user = request.user
+        if user.is_staff:
+            accommodation = AccommodationModel.objects.all()
+            if accommodation.count() > 0:
+                serializer = AccommodationSerializer(accommodation, many=True)
+                return JsonResponse(serializer.data, status=status.HTTP_200_OK, safe=False)
+            else:
+                return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
         else:
-            return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
 
     def post(self, request):
-        serializer = AccommodationSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return JsonResponse(serializer.data, status=status.HTTP_201_CREATED)
-        return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        user = request.user
+        if user.is_staff:
 
+            account = request.user
+            accommodation = AccommodationModel(author=account)
+
+            serializer = AccommodationSerializer(accommodation, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return JsonResponse(serializer.data, status=status.HTTP_201_CREATED)
+            return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
 
 class Accommodation(APIView):
     """
     Retrieve, update or delete a accommodation instance.
     """
+
+    # authentication_classes = [TokenAuthentication]
+    permission_classes = [IsAuthenticated]
 
     schema = AccommodationSchema()
 
@@ -95,27 +115,48 @@ class Accommodation(APIView):
     def get(self, request, pk):
         try:
             accommodation = AccommodationModel.objects.get(pk=pk)
+
             serializer = AccommodationSerializer(accommodation)
+            # update_auth_token(request.user.id)
             return JsonResponse(serializer.data, status=status.HTTP_200_OK)
         except AccommodationModel.DoesNotExist:
+            # update_auth_token(request.user.id)
             return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
 
     def put(self, request, pk):
-        try:
-            accommodation = AccommodationModel.objects.get(pk=pk)
-            serializer = AccommodationSerializer(accommodation, data=request.data)
-            if serializer.is_valid():
-                serializer.save()
-                return JsonResponse(serializer.data, status=status.HTTP_200_OK)
-            else:
-                return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-        except AccommodationModel.DoesNotExist:
-            return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        user = request.user
+        if user.is_staff:
+            try:
+                accommodation = AccommodationModel.objects.get(pk=pk)
+
+                # user = request.user
+                # if accommodation.author != user:
+                #     return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
+
+                serializer = AccommodationSerializer(accommodation, data=request.data)
+                if serializer.is_valid():
+                    serializer.save()
+                    return JsonResponse(serializer.data, status=status.HTTP_200_OK)
+                else:
+                    return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            except AccommodationModel.DoesNotExist:
+                return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        else:
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
 
     def delete(self, request, pk):
-        try:
-            accommodation = AccommodationModel.objects.get(pk=pk)
-            accommodation.delete()
-            return JsonResponse(204, status=status.HTTP_204_NO_CONTENT, safe=False)
-        except AccommodationModel.DoesNotExist:
-            return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        user = request.user
+        if user.is_staff:
+            try:
+                accommodation = AccommodationModel.objects.get(pk=pk)
+
+                # user = request.user
+                # if accommodation.author != user:
+                #     return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
+
+                accommodation.delete()
+                return JsonResponse(204, status=status.HTTP_204_NO_CONTENT, safe=False)
+            except AccommodationModel.DoesNotExist:
+                return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        else:
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
