+++ createAdventure/country/api/views.py
@@ -4,11 +4,10 @@ from requests import Response
 from rest_framework.decorators import api_view
 from rest_framework.parsers import JSONParser
 from rest_framework.views import APIView
-
 from .models import Country as CountryModel
 from .serializer import CountrySerializer
 from rest_framework import generics, status
-
+from rest_framework.permissions import IsAuthenticated, IsAdminUser
 import coreapi
 from rest_framework.schemas import AutoSchema
 
@@ -70,31 +69,50 @@ class listOfCountries(APIView):
     List all countries, or create a new country.
     """
 
+    permission_classes = [IsAuthenticated]
+
     schema = CountrySchema()
 
     def get(self, request):
-        countries = CountryModel.objects.all()
-        if countries.count() > 0:
-            serializer = CountrySerializer(countries, many=True)
-            return JsonResponse(serializer.data, status=status.HTTP_200_OK, safe=False)
+        user = request.user
+        # update_token_view_API(request='PUT', pk=user.id)
+        if user.is_staff:
+            countries = CountryModel.objects.all()
+            if countries.count() > 0:
+                serializer = CountrySerializer(countries, many=True)
+                return JsonResponse(serializer.data, status=status.HTTP_200_OK, safe=False)
+            else:
+                return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
         else:
-            return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
 
     def post(self, request):
-        serializer = CountrySerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return JsonResponse(serializer.data, status=status.HTTP_201_CREATED)
-        return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        user = request.user
+        if user.is_staff:
+
+            account = request.user
+            country = CountryModel(author=account)
+
+            serializer = CountrySerializer(country, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return JsonResponse(serializer.data, status=status.HTTP_201_CREATED)
+            return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
 
 class particularCountry(APIView):
     """
     Retrieve, update or delete a country instance.
     """
 
+    permission_classes = [IsAuthenticated]
+
     schema = CountrySchema()
 
     def get(self, request, pk):
+        # user = request.user
+        # update_token_view_API(request, pk=user.id)
         try:
             country = CountryModel.objects.get(pk=pk)
             serializer = CountrySerializer(country)
@@ -103,21 +121,39 @@ class particularCountry(APIView):
             return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
 
     def put(self, request, pk):
-        try:
-            country = CountryModel.objects.get(pk=pk)
-            serializer = CountrySerializer(country, data=request.data)
-            if serializer.is_valid():
-                serializer.save()
-                return JsonResponse(serializer.data, status=status.HTTP_200_OK)
-            else:
-                return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-        except CountryModel.DoesNotExist:
-            return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        user = request.user
+        if user.is_staff:
+            try:
+                country = CountryModel.objects.get(pk=pk)
+
+                # user = request.user
+                # if country.author != user:
+                #     return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
+
+                serializer = CountrySerializer(country, data=request.data)
+                if serializer.is_valid():
+                    serializer.save()
+                    return JsonResponse(serializer.data, status=status.HTTP_200_OK)
+                else:
+                    return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            except CountryModel.DoesNotExist:
+                return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        else:
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
 
     def delete(self, request, pk):
-        try:
-            country = CountryModel.objects.get(pk=pk)
-            country.delete()
-            return JsonResponse(204, status=status.HTTP_204_NO_CONTENT, safe=False)
-        except CountryModel.DoesNotExist:
-            return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        user = request.user
+        if user.is_staff:
+            try:
+                country = CountryModel.objects.get(pk=pk)
+
+                # user = request.user
+                # if country.author != user:
+                #     return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
+
+                country.delete()
+                return JsonResponse(204, status=status.HTTP_204_NO_CONTENT, safe=False)
+            except CountryModel.DoesNotExist:
+                return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        else:
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
