+++ createAdventure/placeToEat/api/views.py
@@ -4,7 +4,7 @@ from requests import Response
 from rest_framework.decorators import api_view
 from rest_framework.parsers import JSONParser
 from rest_framework.views import APIView
-
+from rest_framework.permissions import IsAuthenticated, IsAdminUser
 from .models import PlaceToEat as PlaceToEatModel
 from .serializer import PlaceToEatSerializer
 from rest_framework import generics, status
@@ -40,27 +40,43 @@ class listOfPlacesToEat(APIView):
     List all placesToEat, or create a new placeToEat.
     """
 
+    permission_classes = [IsAuthenticated]
+
     schema = PlaceToEatSchema()
 
     def get(self, request):
-        placesToEat = PlaceToEatModel.objects.all()
-        if placesToEat.count() > 0:
-            serializer = PlaceToEatSerializer(placesToEat, many=True)
-            return JsonResponse(serializer.data, status=status.HTTP_200_OK, safe=False)
+        user = request.user
+        if user.is_staff:
+            placesToEat = PlaceToEatModel.objects.all()
+            if placesToEat.count() > 0:
+                serializer = PlaceToEatSerializer(placesToEat, many=True)
+                return JsonResponse(serializer.data, status=status.HTTP_200_OK, safe=False)
+            else:
+                return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
         else:
-            return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
 
     def post(self, request):
-        serializer = PlaceToEatSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return JsonResponse(serializer.data, status=status.HTTP_201_CREATED)
-        return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        user = request.user
+        if user.is_staff:
+
+            account = request.user
+            placesToEat = PlaceToEatModel(author=account)
+
+            serializer = PlaceToEatSerializer(placesToEat, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return JsonResponse(serializer.data, status=status.HTTP_201_CREATED)
+            return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
 
 class placeToEat(APIView):
     """
     Retrieve, update or delete a placeToEat instance.
     """
+
+    permission_classes = [IsAuthenticated]
 
     schema = PlaceToEatSchema()
 
@@ -73,21 +89,39 @@ class placeToEat(APIView):
             return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
 
     def put(self, request, pk):
-        try:
-            placeToEat = self.get_object(pk)
-            serializer = PlaceToEatSerializer(placeToEat, data=request.data)
-            if serializer.is_valid():
-                serializer.save()
-                return JsonResponse(serializer.data, status=status.HTTP_200_OK)
-            else:
-                return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-        except PlaceToEatModel.DoesNotExist:
-            return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        user = request.user
+        if user.is_staff:
+            try:
+                placeToEat = self.get_object(pk)
+
+                # user = request.user
+                # if placeToEat.author != user:
+                #     return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
+
+                serializer = PlaceToEatSerializer(placeToEat, data=request.data)
+                if serializer.is_valid():
+                    serializer.save()
+                    return JsonResponse(serializer.data, status=status.HTTP_200_OK)
+                else:
+                    return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            except PlaceToEatModel.DoesNotExist:
+                return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        else:
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
 
     def delete(self, request, pk):
-        try:
-            placeToEat = PlaceToEatModel.objects.get(pk=pk)
-            placeToEat.delete()
-            return JsonResponse(204, status=status.HTTP_204_NO_CONTENT, safe=False)
-        except PlaceToEatModel.DoesNotExist:
-            return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        user = request.user
+        if user.is_staff:
+            try:
+                placeToEat = PlaceToEatModel.objects.get(pk=pk)
+
+                # user = request.user
+                # if placeToEat.author != user:
+                #     return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
+
+                placeToEat.delete()
+                return JsonResponse(204, status=status.HTTP_204_NO_CONTENT, safe=False)
+            except PlaceToEatModel.DoesNotExist:
+                return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        else:
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
