+++ createAdventure/transport/api/views.py
@@ -4,7 +4,7 @@ from requests import Response
 from rest_framework.decorators import api_view
 from rest_framework.parsers import JSONParser
 from rest_framework.views import APIView
-
+from rest_framework.permissions import IsAuthenticated
 from .models import Transport as TransportModel
 from .serializer import TransportSerializer
 from rest_framework import generics, status
@@ -51,28 +51,44 @@ class listOfTransports(APIView):
     """
     List all Transports, or create a new Transport.
     """
+    permission_classes = [IsAuthenticated]
 
     schema = TransportSchema()
 
     def get(self, request):
-        transports = TransportModel.objects.all()
-        if transports.count() > 0:
-            serializer = TransportSerializer(transports, many=True)
-            return JsonResponse(serializer.data, status=status.HTTP_200_OK, safe=False)
+        user = request.user
+        if user.is_staff:
+            transports = TransportModel.objects.all()
+            if transports.count() > 0:
+                serializer = TransportSerializer(transports, many=True)
+                return JsonResponse(serializer.data, status=status.HTTP_200_OK, safe=False)
+            else:
+                return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
         else:
-            return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
 
     def post(self, request):
-        serializer = TransportSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return JsonResponse(serializer.data, status=status.HTTP_201_CREATED)
-        return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        user = request.user
+        if user.is_staff:
+
+            account = request.user
+            transport = TransportModel(author=account)
+
+            serializer = TransportSerializer(transport, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return JsonResponse(serializer.data, status=status.HTTP_201_CREATED)
+            return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+        else:
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
 
 class Transport(APIView):
     """
     Retrieve, update or delete a Transport instance.
     """
+
+    permission_classes = [IsAuthenticated]
 
     schema = TransportSchema()
 
@@ -85,21 +101,39 @@ class Transport(APIView):
             return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
 
     def put(self, request, pk):
-        try:
-            transport = TransportModel.objects.get(pk=pk)
-            serializer = TransportSerializer(transport, data=request.data)
-            if serializer.is_valid():
-                serializer.save()
-                return JsonResponse(serializer.data, status=status.HTTP_200_OK)
-            else:
-                return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-        except TransportModel.DoesNotExist:
-            return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        user = request.user
+        if user.is_staff:
+            try:
+                transport = TransportModel.objects.get(pk=pk)
+
+                # user = request.user
+                # if transport.author != user:
+                #     return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
+
+                serializer = TransportSerializer(transport, data=request.data)
+                if serializer.is_valid():
+                    serializer.save()
+                    return JsonResponse(serializer.data, status=status.HTTP_200_OK)
+                else:
+                    return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            except TransportModel.DoesNotExist:
+                return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        else:
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
 
     def delete(self, request, pk):
-        try:
-            transport = TransportModel.objects.get(pk=pk)
-            transport.delete()
-            return JsonResponse(204, status=status.HTTP_204_NO_CONTENT, safe=False)
-        except TransportModel.DoesNotExist:
-            return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        user = request.user
+        if user.is_staff:
+            try:
+                transport = TransportModel.objects.get(pk=pk)
+
+                # user = request.user
+                # if transport.author != user:
+                #     return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
+
+                transport.delete()
+                return JsonResponse(204, status=status.HTTP_204_NO_CONTENT, safe=False)
+            except TransportModel.DoesNotExist:
+                return JsonResponse(404, status=status.HTTP_404_NOT_FOUND, safe=False)
+        else:
+            return JsonResponse(401, status=status.HTTP_401_UNAUTHORIZED, safe=False)
